requirements.txt
beautifulsoup4==4.6.3
dj-static==0.0.6
Django==1.11.15
django-extensions==2.1.3
django-test-without-migrations==0.6
html5lib==1.0.1
python-decouple==3.1
pytz==2018.5
six==1.11.0
#static3==0.7.0
webencodings==0.5.1
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
kombu==4.2.1
vine==1.1.4

# DRF
certifi==2018.8.24
chardet==3.0.4
djangorestframework==3.8.2
html5lib==1.0.1
httpie==0.9.9
idna==2.7
Pygments==2.2.0
requests==2.19.1
urllib3==1.23
webencodings==0.5.1

